[BE] find: search 로직 오류 발견
user like 하면서 덮어씌워진 search 로직 발견
