+++ frontend/components/SearchSamples/SearchAlbumList/index.tsx
@@ -1,0 +1,42 @@
+import React, { FC } from "react";
+import StyledAlbumList from "./styled";
+import SearchAlbumCard from "./SearchAlbumCard";
+
+interface Album {
+  id: number;
+  albumName: string;
+  description: string;
+  cover: string;
+  artistId: number;
+  Artists: {
+    artistName: string;
+  };
+}
+
+const SearchAlbumList = ({ data }: { data: Album[] }): React.ReactElement => {
+  return (
+    <StyledAlbumList>
+      {data.map((album) => {
+        const {
+          albumName,
+          cover,
+          Artists: { artistName },
+          id,
+          artistId,
+        } = album;
+        return (
+          <SearchAlbumCard
+            key={id}
+            albumName={albumName}
+            cover={cover}
+            artistName={artistName}
+            id={id}
+            artistId={artistId}
+          />
+        );
+      })}
+    </StyledAlbumList>
+  );
+};
+
+export default SearchAlbumList;
