+++ frontend/components/SearchSamples/SearchAlbumList/styled.ts
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+const StyledAlbums = styled.div`
+  width: 100%;
+  display: flex;
+  flex-wrap: wrap;
+`;
+
+export default StyledAlbums;
